% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregions2-package.R
\name{gaz_rest}
\alias{gaz_rest}
\title{Marine Regions Gazetteer RESTful services}
\description{
\emph{RESTful service}
REST (REpresentational State Transfer) is a simple stateless architecture that generally runs over HTTP.

\link{mregions2} makes use of the RESTful services created and maintained by Marine Regions. You can find more
information at the overview of the \href{https://marineregions.org/gazetteer.php?p=webservices}{Marine Regions Gazetteer web services}
}
\examples{
\dontrun{}
}
