% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gaz_sources.R
\name{gaz_rest_records_by_source}
\alias{gaz_rest_records_by_source}
\title{Retrieve Gazetteer Records by Source}
\usage{
gaz_rest_records_by_source(source, with_geometry = FALSE)
}
\arguments{
\item{source}{(character) A source from \code{\link[=gaz_rest_sources]{gaz_rest_sources()}}}

\item{with_geometry}{(logical) Add geometries to the result data frame? Default = FALSE}
}
\value{
A data frame with Gazetteer entries
}
\description{
Retrieve Gazetteer Records by Source
}
\examples{
\donttest{
gaz_rest_records_by_source("ICES Ecoregions")
}
}
\seealso{
\link{gaz_rest}
}
