% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_gaz_types.R
\name{gaz_rest_records_by_type}
\alias{gaz_rest_records_by_type}
\title{Retrieve Gazetteer Records by Placetype}
\usage{
gaz_rest_records_by_type(type, with_geometry = FALSE)
}
\arguments{
\item{type}{(character) The placetype from \code{\link[=gaz_rest_types]{gaz_rest_types()}}}

\item{with_geometry}{(logical) Add geometries to the result data frame? Default = FALSE}
}
\value{
A data frame with Gazetteer entries
}
\description{
Retrieve Gazetteer Records by Placetype
}
\examples{
\donttest{
gaz_rest_records_by_type("FAO Subdivisions")
gaz_rest_records_by_type("EEZ")
}
}
\seealso{
\link{gaz_rest}, \code{\link[=gaz_rest_types]{gaz_rest_types()}}
}
