% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gaz_sources.R
\name{gaz_rest_source_by_sourceid}
\alias{gaz_rest_source_by_sourceid}
\title{Get the name of a source by providing a sourceID}
\usage{
gaz_rest_source_by_sourceid(sourceid)
}
\arguments{
\item{sourceid}{(integer) A valid sourceID}
}
\value{
a named vector with the source name and, if available, the url to the source.
}
\description{
Get the name of a source by providing a sourceID
}
\examples{
\donttest{
gaz_rest_source_by_sourceid(390)
gaz_rest_source_by_sourceid(657)
}
}
\seealso{
\link{gaz_rest}, \code{\link[=gaz_sources]{gaz_sources()}}
}
