% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gaz_sources.R
\name{gaz_rest_sources}
\alias{gaz_rest_sources}
\title{Get all the Marine Regions sources}
\usage{
gaz_rest_sources()
}
\value{
a data frame with three columns:
\itemize{
\item \code{sourceID}: the identifier of the source in the Marine Regions Gazetteer database.
\item \code{source}: the name of the source.
\item \code{sourceURL}: if available, the URL of the source.
}
}
\description{
Get all the Marine Regions sources
}
\details{
gaz_search() is a memoised function from gaz_rest_search(). See \code{\link[memoise:memoise]{memoise::memoise()}}.
}
\examples{
\donttest{
# This
gaz_rest_sources()

# is the same as
gaz_sources()
}
}
\seealso{
\link{gaz_rest}, \code{\link[=gaz_search_by_source]{gaz_search_by_source()}}, \code{\link[=gaz_rest_records_by_source]{gaz_rest_records_by_source()}}, \code{\link[=gaz_rest_source_by_sourceid]{gaz_rest_source_by_sourceid()}}
}
