% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_gaz_types.R
\name{gaz_search_by_type}
\alias{gaz_search_by_type}
\title{Retrieve Gazetteer Records by Placetype}
\usage{
gaz_search_by_type(type, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=gaz_rest_records_by_type]{gaz_rest_records_by_type}}
  \describe{
    \item{\code{with_geometry}}{(logical) Add geometries to the result data frame? Default = FALSE}
  }}
}
\value{
A data frame with Gazetteer entries
}
\description{
Retrieve Gazetteer Records by Placetype
}
\examples{
gaz_search_by_type("FAO Subdivisions")
gaz_search_by_type("EEZ")
}
\seealso{
\link{gaz_rest}
}
