% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_gaz_types.R
\name{gaz_types}
\alias{gaz_types}
\title{Get all the place types of the Marine Regions Gazetteer}
\usage{
gaz_types()
}
\value{
a data frame with three columns:
\itemize{
\item \code{typeID}: the identifier of the place type in the Marine Regions Gazetteer database.
\item \code{type}: the name of the place type.
\item \code{description}: if available, the description of the place type.
}
}
\description{
Get all the place types of the Marine Regions Gazetteer
}
\examples{
\donttest{
# This
gaz_rest_types()

# is the same as
gaz_types()
}
}
\seealso{
\link{gaz_rest}
}
