% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_dev.R
\name{marine_region}
\alias{marine_region}
\title{R6 Class Representing a Marine Region Object}
\description{
A ´MarineRegion` object corresponds to a record in the Marine Gazetteer, available at https://marineregions.org.
It contains relevant information such as the Name, the MRGID (Marine Regions Geographic IDentifier), the Placetype, the Source,...
Inititalize `Marineregion` object
}
\examples{

## ------------------------------------------------
## Method `marine_region$print`
## ------------------------------------------------

guadelupe <- mr_marine_reggion(name = "Guadelupe")
guadelupe
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{marine_region$new()}}
\item \href{#method-print}{\code{marine_region$print()}}
\item \href{#method-get_info}{\code{marine_region$get_info()}}
\item \href{#method-get_geometry}{\code{marine_region$get_geometry()}}
\item \href{#method-get_geometry_info}{\code{marine_region$get_geometry_info()}}
\item \href{#method-clone}{\code{marine_region$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This method gets called whenever a new `MarineRegion` object is created. It created all necessary attributes and e.g. includes that the object is immediately printed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$new(add_geometry, mrgid, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_geometry}}{specify if the geometry is wanted or not}

\item{\code{mrgid}}{the Marine Regions Geographic IDentifier}

\item{\code{name}}{the name of the `MarineRegion` object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an initialized `MarineRegion` object
Print important attributes of the class `MarineRegion`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
important attributes of the class `MarineRegion`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{guadelupe <- mr_marine_reggion(name = "Guadelupe")
guadelupe
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_info}{}}}
\subsection{Method \code{get_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$get_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_geometry"></a>}}
\if{latex}{\out{\hypertarget{method-get_geometry}{}}}
\subsection{Method \code{get_geometry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$get_geometry()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_geometry_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_geometry_info}{}}}
\subsection{Method \code{get_geometry_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$get_geometry_info(geometry)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
