% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_add_geometry}
\alias{mr_add_geometry}
\title{Add geometry to a data frame containing a column named MRGID}
\usage{
mr_add_geometry(x)
}
\arguments{
\item{x}{a data frame obtained via mr_gaz_record(s)_by_*()}
}
\value{
the same data frame with a new column the_geom
}
\description{
Add geometry to a data frame containing a column named MRGID
}
\examples{
mr_gaz_record_by_mrgid(19518, with_geometry = TRUE)

# It's the same as
mr_gaz_record_by_mrgid(19518) \%>\% mr_add_geometry()

mr_gaz_records_by_name("Belgium") \%>\% mr_add_geometry()
}
