% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_mr_gaz_geom.R
\name{mr_gaz_geometry}
\alias{mr_gaz_geometry}
\alias{mr_gaz_geometries}
\title{Funtions to return the geometry of a marine regions geoobject}
\usage{
mr_gaz_geometry(mrgid, sourceid, attribute_value, format = "sfc", ...)

mr_gaz_geometries(mrgid, format = "sfc", multipart = TRUE, ...)
}
\arguments{
\item{mrgid}{int. A valid marine regions gazetteer identifier}

\item{sourceid}{int. A source ID}

\item{attribute_value}{char. The attribute value that identifies the geoobject}

\item{format}{The preferred output format. One of:
- "sfc": Simple Feature geometry object. See 'sf'
- "wkt": Geometry representation as Well-Known Text
- "rdf": Geometry as an object of class 'rdf". See 'rdflib'

Default is "sfc"}
}
\value{
the geometry of a marine regions geoobject
}
\description{
Funtions to return the geometry of a marine regions geoobject
}
\examples{
mr_gaz_geometry(mrgid = 3293, sourceid = 79, attribute_value = 3293)
mr_gaz_geometries(mrgid = 58, format = "wkt")
}
