% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_records.R
\name{mr_gaz_info}
\alias{mr_gaz_info}
\title{Retrieve information of Marine Gazetteer Records}
\usage{
mr_gaz_info(info = c("sources", "placetypes"))
}
\arguments{
\item{info}{The information wished to retrieve. Must be one of: `c("sources", "placetypes")`.}
}
\value{
Sources: A tibble with the source and source URL.
Placetypes: A tibble with the ID, name and description of each Placetype available in the Marine Gazetteer \url{https://marineregions.org/gazetteer.php}.
}
\description{
Available Placetypes and Sources of the Marine Gazetteer Records, available at: \url{https://marineregions.org/gazetteer.php}.
Useful for retrieving information about Gazetteer Records based on Placetypes \code{\link{mr_gaz_records_by_type}} and Sources \code{\link{mr_gaz_records_by_source}}.
}
\examples{
placetypes <- mr_gaz_info(info = "placetypes")
placetypes$type[28]
# [1] "Sandbank"

sources <- mr_gaz_info(info = "sources")
sources$source[36]
# [1] "GeoNames"
}
