% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_record.R
\name{mr_gaz_ldes}
\alias{mr_gaz_ldes}
\title{Retrieves a Marine Regions Gazetteer record as Open Linked Data}
\usage{
mr_gaz_ldes(mrgid, type = "list")
}
\arguments{
\item{mrgid}{Marine Regions Gazetter Unique Identifier}

\item{type}{Return resuts as RDF or as list}
}
\value{
A Marine Regions Gazetteer record as triples in RDF or list
}
\description{
Retrieves a Marine Regions Gazetteer record as Open Linked Data
}
\examples{
test <- mr_gaz_ldes(26567)
test <- mr_gaz_ldes(3293, type = "list")
test <- mr_gaz_ldes(1902, type = "rdf")
}
