% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_record.R
\name{mr_gaz_record}
\alias{mr_gaz_record}
\title{Get a Marine Regions Gazetteer record}
\usage{
mr_gaz_record(mrgid, add_geometry = TRUE)
}
\arguments{
\item{mrgid}{Marine Regions Gazetter Unique Identifier}

\item{add_geometry}{set FALSE to not retrieve the geometry}
}
\value{
A Marine Regions Gazetteer record
}
\description{
Get a Marine Regions Gazetteer record
}
\examples{
test <- mr_gaz_record(3293)
test <- mr_gaz_record(4280)
}
