% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_mr_gaz_records.R
\name{mr_gaz_record_by_mrgid}
\alias{mr_gaz_record_by_mrgid}
\title{Get one record for the given MRGID}
\usage{
mr_gaz_record_by_mrgid(mrgid, with_geometry = FALSE, rdf = FALSE)
}
\arguments{
\item{mrgid}{An existing Marine Regions Gazetteer Identifier}

\item{with_geometry}{Logical. Add geometry to the result data frame? Default = FALSE}

\item{rdf}{Logical. Return an object of class "rdf". See package 'rdflib'}
}
\description{
Get one record for the given MRGID
}
