% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_mr_gaz_records.R
\name{mr_gaz_records_by_lat_long}
\alias{mr_gaz_records_by_lat_long}
\title{Get all gazetteer records where the geometry intersects with the given latitude and longitude}
\usage{
mr_gaz_records_by_lat_long(
  latitude,
  longitude,
  with_geometries = FALSE,
  typeid = NULL
)
}
\arguments{
\item{latitude}{A decimal number which ranges from -90 to 90. Coordinates are assumed to be in WGS84}

\item{longitude}{A decimal number which ranges from -180 to 180. Coordinates are assumed to be in WGS84}

\item{with_geometries}{Logical. Add geometries to the result data frame? Default = FALSE}

\item{typeid}{Restrict to one or more placetypeIDs. See function mr_gaz_types() to retrieve a list of placetypeIDs. Default = NULL}
}
\description{
Get all gazetteer records where the geometry intersects with the given latitude and longitude
}
\examples{
mr_gaz_records_by_lat_long(51.21551, 2.927)
mr_gaz_records_by_lat_long(51.21551, 2.927, with_geometries = TRUE, typeid = c(255, 259))
}
