% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_records.R
\name{mr_gaz_records_by_latlon}
\alias{mr_gaz_records_by_latlon}
\title{Retrieve Gazetteer Records by Lat-Lon Coordinates}
\usage{
mr_gaz_records_by_latlon(lat, lon, lat_radius = 0, lon_radius = 0)
}
\arguments{
\item{lat}{A decimal number which ranges from -90 to 90.}

\item{lon}{A decimal number which ranges from -180 to +180.}

\item{lat_radius}{Searches for latitudes in the range 'latitude'-'latRadius' and 'latitude'+'latRadius' default = 0.}

\item{lon_radius}{Searches for longitudes in the range 'longitude'-'longRadius' and 'longitude'+'longRadius' default = 0.}
}
\value{
A tibble with the first 100 Gazetteer records of the specified coordinates.
}
\description{
Get the first 100 [Marine Gazetteer](https://marineregions.org/gazetteer.php) records, where their centroid is within the bounding box calculated by latitude (+/- radius) and longitude (+/- radius)
}
\examples{
oostende_lat <- 51.21551
oostende_lon <- 2.927
oostende <- mr_gaz_records_by_latlon(ostende_lat, ostende_lon, lat_radius = 0, lon_radius = 0)

some_atlantic_lat <- -37
some_atlantic_lon <- 37
some_atlantic_location <- mr_gaz_records_by_latlon(some_atlantic_lat, some_atlantic_lon)
}
