% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_records.R
\name{mr_gaz_records_by_names}
\alias{mr_gaz_records_by_names}
\title{Retrieve Gazetteer Records by a list of Names}
\usage{
mr_gaz_records_by_names(names, like = TRUE, fuzzy = FALSE)
}
\arguments{
\item{names}{Names that are retrieved from the}

\item{like}{Formats the name into a SQL-like syntax. Default = TRUE.}

\item{fuzzy}{Uses Levenshtein query to find nearest matches. Default = FALSE.}
}
\value{
Tibble with Gazetteer records.
}
\description{
Retrieve all Records of the [Marine Gazetteer](https://marineregions.org/gazetteer.php) that contain one or more parameters in `names` in their `preferredGazetteerName`.
}
\examples{
names <- c("Atlantic Ocean", "Vanuatu", "Mariana Trench")
result <- mr_gaz_records_by_names(names)
}
