% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_mr_gaz_types.R
\name{mr_gaz_records_by_type}
\alias{mr_gaz_records_by_type}
\title{Retrieve Gazetteer Records by Placetype}
\usage{
mr_gaz_records_by_type(placetype, with_geometries = FALSE)
}
\arguments{
\item{with_geometries}{Logical. Add geometries to the result data frame? Default = FALSE}

\item{type}{The placetype from mr_gaz_types()}
}
\value{
A tibble with all Gazetteer records of the specified placetype.
}
\description{
Retrieve Gazetteer Records by Placetype
}
\examples{
mr_gaz_records_by_type("FAO Subdivisions")
mr_gaz_records_by_type("EEZ")
}
