% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_records.R
\name{mr_gaz_relations_full_by_MRGID}
\alias{mr_gaz_relations_full_by_MRGID}
\title{Get all Relations for a Marine Gazetteer Record}
\usage{
mr_gaz_relations_full_by_MRGID(mrgid)
}
\arguments{
\item{mrgid}{The Marine Regions Geographic IDentifier.}
}
\value{
A `tibble` with all records that are related. The column `relation` contains the type of relation.
Explanations of the relation types can be found at: https://marineregions.org/ontology/documentation.html#objectproperties.
}
\description{
Get all Relations for a Marine Gazetteer Record
}
\examples{
belgium_mrgid <- 14
belgium_relations <- mr_gaz_relations_full_by_MRGID(belgium_mrgid)
}
