% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_mr_gaz_sources.R
\name{mr_gaz_source_by_sourceid}
\alias{mr_gaz_source_by_sourceid}
\title{Get the name of a source by providing a sourceID}
\usage{
mr_gaz_source_by_sourceid(sourceid)
}
\arguments{
\item{sourceid}{int. A valid sourceID}
}
\value{
a named vector with the source name and, if available, the url to the source.
}
\description{
Get the name of a source by providing a sourceID
}
\examples{
mr_gaz_source_by_sourceid(390)
mr_gaz_source_by_sourceid(657)
}
