% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_mr_gaz_sources.R
\name{mr_gaz_sources}
\alias{mr_gaz_sources}
\title{Get all the Marine Regions sources}
\usage{
mr_gaz_sources()
}
\value{
a tibble with three columns:
- sourceID: the identifier of the source in the database.
- source: the name of the source
- sourceURL: if available, the URL of the source.
}
\description{
Get all the Marine Regions sources
}
\examples{
mr_gaz_sources()
}
