% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_mr_gaz_types.R
\name{mr_gaz_types}
\alias{mr_gaz_types}
\title{Get all the placetypes of the Marine Regions Gazetteer}
\usage{
mr_gaz_types()
}
\value{
a tibble with the type and definition if available
}
\description{
Get all the placetypes of the Marine Regions Gazetteer
}
\examples{
types <- mr_gaz_types()
}
