% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_has_geometry}
\alias{mr_has_geometry}
\title{Check if marine region has a geometry}
\usage{
mr_has_geometry(mrgid)
}
\arguments{
\item{mrgid}{The Marine Regions Geographic IDentifier.}
}
\value{
TRUE if the a geometry is available, otherwise FALSE.
}
\description{
Check if marine region has a geometry
}
\examples{
guadelupe_mrgid <- mr_gaz_records_by_name(name = "Guadelupe", count = 1)$MRGID
guadelupe_hasgeometry <- mr_has_geometry(guadelupe_mrgid)

belgium_mrgid <- mr_gaz_records_by_name(name = "Belgium", count = 1)$MRGID
belgium_hasgeometry <- mr_has_geometry(belgium_mrgid)
}
