% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_http_error_message}
\alias{mr_http_error_message}
\title{Retrieve error message for frequent http errors}
\usage{
mr_http_error_message(response)
}
\arguments{
\item{response}{httr2 response object.}
}
\value{
error message if the response status code is one of the status codes handled by the function.
}
\description{
Retrieve error message for frequent http errors
}
\examples{
failresp <- httr2::response(status_code = 404)
successresp <- httr2::response(status_code = 200)
mr_http_error_message(successresp)
mr_http_error_message(failresp)
}
