% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_dev.R
\name{mr_marine_region}
\alias{mr_marine_region}
\title{Get a marine region}
\usage{
mr_marine_region(name = NA, mrgid = NA, add_geometry = TRUE)
}
\arguments{
\item{name}{The `preferredGazetteerName` of a marine gazetteer record. Can be retrieved with `mr_gaz_records_by_name()`.}

\item{mrgid}{The `Marine Regions Geographical IDentifier`.}

\item{add_geometry}{Statement if geospatial data should be retrieved.}
}
\value{
An instance of the class `MarineRegion`.
}
\description{
If only a name is given, the MRGID from the first record found by `mr_gaz_record_by_name(name)` will be fetched. Since names are not unique, prefer to input MRGIDs.
}
\examples{
german_part_north_sea <- mr_marine_region(name = "German Part of the North Sea")
high_seas_mrgid <- mregions2::mr_gaz_records_by_name("High Seas", count = 1)$MRGID
high_seas <- mr_marine_region(mrgid = high_seas_mrgid, add_geometry = FALSE)
}
