% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_dev.R
\name{mr_marine_region_extend}
\alias{mr_marine_region_extend}
\title{Add elements to a marine region}
\usage{
mr_marine_region_extend(region, extension_name, extension_value)
}
\arguments{
\item{region}{Must be an object of the class `MarineRegion`}

\item{extension_name}{Must be of class character.}

\item{extension_value}{Can be a value or a function.}
}
\value{
Object that inherits of class `MarineRegion` that contains the new element specified.
}
\description{
Add elements to a marine region
}
\examples{
mexico <- mr_marine_region(mrgid = 2224)
children <- function(mrgid) mregions2::mr_gaz_relations_by_MRGID(mrgid, direction = "lower", type = "all")
mr_marine_region_extend(region = mexico,  extension_name = "children", extension_value = children)
}
