% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_null_to_na}
\alias{mr_null_to_na}
\title{Replace NULL values with NA in list}
\usage{
mr_null_to_na(list_with_NULL)
}
\arguments{
\item{list_with_NULL}{list element passed to be changed to NA if NULL.}
}
\value{
elements in e.g. a list with NA instead of NULL.
}
\description{
useful when running `tibble::as_tibble()` which does not accept NULL.
}
\examples{
y <- list(
one = c(month.abb, 5),
two = c(NULL, 5, NA),
three = NULL,
four = list(NULL) # does not change nested NULL for now
)
z <- mr_null_to_na(y)
}
