% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_resp_to_tibble}
\alias{mr_resp_to_tibble}
\title{Transform httr2 response into tibble}
\usage{
mr_resp_to_tibble(resp, unpack = FALSE)
}
\arguments{
\item{resp}{response from httr2 request.}

\item{unpack}{Set `TRUE` when `mr_gaz_records_by_names()` is run. This webservice nests the result once more.}
}
\value{
A tibble containing the json response body from the httr2 response.
}
\description{
Transform httr2 response into tibble
}
\examples{
name <- "High Sea"
count <- 5
url <- "https://marineregions.org//rest/getGazetteerRecordsByName.json/"

req <- httr2::request(url) \%>\%
 httr2::req_url_path_append(utils::URLencode(name)) \%>\%
 httr2::req_url_path_append("/")

resp <- req \%>\%
 httr2::req_url_query(
  `like` = TRUE,
  `fuzzy` = FALSE,
  `offset` = 0,
  `count` = count) \%>\%
 httr2::req_perform()

res <- mr_resp_to_tibble(resp)
}
