% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregions2-package.R
\docType{package}
\name{mregions2-package}
\alias{mregions2}
\alias{mregions2-package}
\title{mregions2: Marine Regions Data from 'Marineregions.org'}
\description{
Retrieve marine geospatial information from the Marineregions gazetteer. Access and handle data products from the Flanders Marine Institute (VLIZ) geoserver.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lifewatch/mregions2}
  \item \url{https://lifewatch.github.io/mregions2/}
  \item Report bugs at \url{https://github.com/lifewatch/mregions2/issues}
}

}
\author{
\strong{Maintainer}: Salvador Fernandez \email{salvador.fernandez@vliz.be} (\href{https://orcid.org/0000-0003-0535-7677}{ORCID}) (salvafern)

Authors:
\itemize{
  \item Lotte Pohl \email{lotte.pohl@imbrsea.eu} (\href{https://orcid.org/0000-0002-7607-7018}{ORCID}) (whaleshark99)
}

Other contributors:
\itemize{
  \item LifeWatch Belgium (https://lifewatch.be) [funder]
}

}
\keyword{internal}
