% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_mrp_get.R
\name{mrp_get}
\alias{mrp_get}
\title{Get a data product}
\usage{
mrp_get(product_name, ...)
}
\arguments{
\item{product_name}{(character) Identifier of the data product. See \code{\link[=mrp_list]{mrp_list()}}}

\item{...}{Any \href{https://docs.geoserver.org/latest/en/user/services/wfs/vendor.html}{vendor params} to be passed to the \href{https://docs.geoserver.org/latest/en/user/services/wfs/reference.html?highlight=getfeatures#getfeature}{GetFeature} request. See details.}
}
\value{
An sf object with the Marine Regions data product
}
\description{
Get a data product
}
\details{
This function uses \href{https://en.wikipedia.org/wiki/Web_Feature_Service}{WFS services} to load retrieve the Marine Regions Data products. The data products
are stored in a \href{https://geo.vliz.be}{geoserver instance} hosted and maintained by the \href{https://www.vliz.be}{Flanders Marine Institute (VLIZ)}
}
\section{Vendor Params}{
\href{https://docs.geoserver.org/latest/en/user/services/wfs/vendor.html}{Vendor params} can be passed to a \href{https://docs.geoserver.org/latest/en/user/services/wfs/reference.html?highlight=getfeatures#getfeature}{WFS GetFeature} request.
\code{mregions2} uses the \href{https://github.com/eblondel/ows4R}{ows4R} package as the \href{https://github.com/eblondel/ows4R/wiki#wfs}{interface to query the WFS service with R.} You can pass any vendor parameters via \code{...}.
For instance, you can pass CQL or OGC filters with \code{cql_filter="<the filter>"} or \code{filter="<the filter>"}. You can change the projection with \code{srsName="<CRS>"}, only the features in a certain bounding box with \code{bbox="a1,b1,a2,b2,<CRS>"}
where \code{a1}, \code{b1}, \code{a2} and \code{b2} are the coordinates and the \code{CRS} selects the coordinate reference system of this bounding box, or \href{https://docs.geoserver.org/latest/en/user/services/wfs/vendor.html}{any other vendor parameters accepted by geoserver}.
}

\section{Filtering}{
Both the \href{https://portal.ogc.org/files/96288}{Contextual Query Language (CQL) filter} and the \href{https://www.ogc.org/standards/filter}{standard OGC filter specification} allow to
query the server before performing a request. This will boost performance as you will only retrieve the area of your interest. It is possible to query on attributes, but also perform
geospatial queries. For instance, you can query a bounding box of interest.

A tutorial on CQL filters is available in the \href{https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html}{geoserver web site}.
}

\examples{
\dontrun{
# See the list of all data products
View(mrp_list())

# We want the Exclusive Economic Zones of Portugal. Let's first visualize the product:
mrp_view("eez")

# Let's see all the columns on this data product
mrp_colnames("eez")

# We should query on sovereign. Let's see all the possible values of sovereign1, sovereign2 and sovereign3
sov1 = mrp_col_unique("eez", "sovereign1")
sov2 = mrp_col_unique("eez", "sovereign2")
sov3 = mrp_col_unique("eez", "sovereign3")

# Is Portugal a value in the sovereign1, 2 and 3?
"Portugal" \%in\% sov1
#> [1] TRUE

"Portugal" \%in\% sov2
#> [1] FALSE

"Portugal" \%in\% sov3
#> [1] FALSE

# Portugal is only in sovereign1. Let's write a CQL filter to get only the EEZs of Portugal, or those
# where Portugal is a party of a dispute or a joint regime
portugal_eez <- mrp_get("eez", cql_filter = "sovereign1 = 'Portugal'")

# You can also limit the number of features to be requested
mrp_get("eez", count = 5)
}
}
\seealso{
\code{\link[=mrp_list]{mrp_list()}} to describe the list of products, \code{\link[=mrp_view]{mrp_view()}} to visualize the data product in advance, \code{\link[=mrp_colnames]{mrp_colnames()}} and \code{\link[=mrp_col_unique]{mrp_col_unique()}} to get the name, data type and unique values of a the columns of a data product, useful to query
with the arguments \code{cql_filter} or \code{filter}
}
