% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_mrp_list.R
\name{mrp_list}
\alias{mrp_list}
\title{Available data products at Marine Regions}
\usage{
mrp_list()
}
\description{
A function returning data frame including the name, abstract and
some other relevant about each data product in Marine Regions.
\itemize{
\item \code{title}: name of the data products
\item \code{data_product}: identifier of the data product. Use in \code{\link[=mrp_get]{mrp_get()}}
\item \code{license}: terms of use of the data products
\item \code{citation}: preferred citation of the data products
\item \code{doi}: ISO 26324 \href{https://doi.org}{Digital Object Identifier}
\item \code{imis}: url of the data products in the \href{https://vliz.be/imis?}{Integrated Marine Information System (IMIS)}
\item \code{id}: geoserver ID, reserved for internal use. Format 'namespace:data_product'
}
}
\references{
\itemize{
\item Sources: \url{https://marineregions.org/sources.php}
\item Methodology of the creation of the Maritime Boundaries: \url{https://marineregions.org/eezmethodology.php}
\item Direct download: \url{https://marineregions.org/downloads.php}
}
}
\seealso{
\code{\link[=mrp_view]{mrp_view()}}, \code{\link[=mrp_colnames]{mrp_colnames()}}, \code{\link[=mrp_get]{mrp_get()}}
}
