% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrp_ontology}
\alias{mrp_ontology}
\title{Marine Regions Data Products Ontology}
\format{
\subsection{\code{mrp_ontology}}{

A data frame with 374 rows and 4 columns:
\describe{
\item{layer}{Identifier of the data product. Use in \code{\link[=mrp_get]{mrp_get()}}}
\item{colname}{Name of the columns of each data product.}
\item{type}{Data type of the column.}
\item{definition}{Definition of the column.}
}
}
}
\source{
\url{https://marineregions.org/sources.php} \url{https://marineregions.org/eezattribute.php}
}
\usage{
mrp_ontology
}
\description{
More information available at \code{vignette("mrp_ontology", package = "mregions2")}
}
\examples{
mrp_ontology
}
\keyword{datasets}
