% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{req_mr_user_agent}
\alias{req_mr_user_agent}
\title{Get user agent and package version}
\usage{
req_mr_user_agent(.)
}
\arguments{
\item{.}{A http request.}
}
\value{
A character string that unites the user agent of the http request and the package version.
}
\description{
Get user agent and package version
}
\examples{
request("http://example.com") \%>\% req_mr_user_agent() \%>\% req_dry_run()
}
