% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{req_URL}
\alias{req_URL}
\title{Create a base URL for a http request.}
\usage{
req_URL(api_type, file_format, method)
}
\arguments{
\item{api_type}{Type of API architecture. Must be one of c("rest", "soap").}

\item{file_format}{File format. Must be one of c("json", "xml", "ttl", "jsonld").}

\item{method}{RESTful method. Check the Marineregions gazetteer at https://marineregions.org/gazetteer.php?p=webservices for available methods.}
}
\value{
A base URL to append queries to.
}
\description{
Create a base URL for a http request.
}
\examples{
api <- "rest"
file <- "json"
method <- "getGazetteerRecordsByName"
req_URL(api_type = api, file_format = file, method = method)
# https://marineregions.org//rest/getGazetteerRecordsByName.json/
}
