% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_dev.R
\name{marine_region}
\alias{marine_region}
\title{R6 Class Representing a Marine Region Object}
\description{
A ´MarineRegion` object corresponds to a record in the Marine Gazetteer, available at \url{https://marineregions.org}.
It contains relevant information such as the Name, the MRGID (Marine Regions Geographic IDentifier), the Placetype, the Source,...
}
\examples{

## ------------------------------------------------
## Method `marine_region$print`
## ------------------------------------------------

guadelupe <- mr_marine_region(name = "Guadelupe")
guadelupe
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{add_geometry}}{TRUE/FALSE statement if a `MarineRegion`'s geometry should be added or not.}

\item{\code{mrgid}}{Marine Regions Geographic IDentifier.}

\item{\code{name}}{`preferredGazetteerName` of the `MarineRegion`.}

\item{\code{info}}{Information about the `MarineRegion`.}

\item{\code{geometry}}{Geometry of the `MarineRegion`.}

\item{\code{source}}{Source of the `MarineRegion`.}

\item{\code{longitude}}{Longitude of the `MarineRegion`.}

\item{\code{latitude}}{Latitude of the `MarineRegion`.}

\item{\code{status}}{Status of the `MarineRegion`. `status == "standard"` should be preferred.}

\item{\code{placetype}}{Placetype of the `MarineRegion`.}

\item{\code{relations}}{Related Regions of the `MarineRegion`.}

\item{\code{area}}{Area in `km^2` of the `MarineRegion`.
Inititalize `Marineregion` object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{marine_region$new()}}
\item \href{#method-print}{\code{marine_region$print()}}
\item \href{#method-clone}{\code{marine_region$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This method gets called whenever a new `MarineRegion` object is created. It created all necessary attributes and e.g. includes that the object is immediately printed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$new(add_geometry, mrgid, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_geometry}}{specify if the geometry is wanted or not}

\item{\code{mrgid}}{the Marine Regions Geographic IDentifier}

\item{\code{name}}{the name of the `MarineRegion` object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an initialized `MarineRegion` object
Print important attributes of the class `MarineRegion`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
if an instance of `MarineRegion` is called, the `print` function will be run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
important attributes of the class `MarineRegion`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{guadelupe <- mr_marine_region(name = "Guadelupe")
guadelupe
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{marine_region$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
