% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_record.R
\name{mr_gaz_geometry}
\alias{mr_gaz_geometry}
\title{Get the geometry a Marine Regions Gazetteer record as Well Known Text}
\usage{
mr_gaz_geometry(mrgid)
}
\arguments{
\item{mrgid}{Marine Regions Gazetter Unique Identifier}
}
\value{
A string with the geometry of the Marine Regions object as WKT
}
\description{
Get the geometry a Marine Regions Gazetteer record as Well Known Text
}
\examples{
test <- mr_gaz_geometry(4280)
test <- mr_gaz_geometry(3293)
test <- mr_gaz_geometry(26567)
}
