% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_records.R
\name{mr_gaz_records_by_name}
\alias{mr_gaz_records_by_name}
\title{Retrieve Gazetteer Records by Name}
\usage{
mr_gaz_records_by_name(
  name,
  count = 100,
  like = TRUE,
  fuzzy = FALSE,
  offset = 0
)
}
\arguments{
\item{name}{GazetteerName of the marine region of interest. Search names here: https://marineregions.org/gazetteer.php?p=search}

\item{count}{Number of records to retrieve.}

\item{like}{Adds a '%'-sign before and after the GazetteerName.}

\item{fuzzy}{Uses Levenshtein query to find nearest matches.}

\item{offset}{Start record number, in order to page through next batch of results.}
}
\value{
A tibble with Gazetter records.
}
\description{
Retrieve Gazetteer Records by Name
}
\examples{
mr_gaz_records_by_name(name = "High Seas", count = 5)
}
