% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_mr_gaz_records.R
\name{mr_gaz_records_by_name}
\alias{mr_gaz_records_by_name}
\title{Get Gazetteer Records for a given name}
\usage{
mr_gaz_records_by_name(
  name,
  with_geometries = FALSE,
  typeid = NULL,
  language = NULL,
  like = TRUE,
  fuzzy = FALSE
)
}
\arguments{
\item{name}{Term to search in the Marine Regions Gazetteer}

\item{with_geometries}{Logical. Add geometries to the result data frame? Default = FALSE}

\item{typeid}{Restrict to one or more placetypeIDs. See function mr_gaz_types() to retrieve a list of placetypeIDs. Default = NULL}

\item{language}{Restrict to one language. Provide as a 2 digits ISO-639. Default = NULL}

\item{like}{Logical. Add a '%'-sign before and after the name? (SQL LIKE function). Default = TRUE}

\item{fuzzy}{Logical. Use Levenshtein query to find nearest matches? Default = FALSE}
}
\description{
Get Gazetteer Records for a given name
}
\examples{
mr_gaz_records_by_name("Belgian Exclusive Economic Zone", with_geometry = TRUE)
mr_gaz_records_by_name("Bélgica", language = "es")
mr_gaz_records_by_name("Belgica", language = "es", fuzzy = TRUE)
mr_gaz_records_by_name("Belgium", typeid = c(350, 351))
}
