% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_mr_gaz_records.R
\name{mr_gaz_records_by_names}
\alias{mr_gaz_records_by_names}
\title{Get Gazetteer Records for all given names}
\usage{
mr_gaz_records_by_names(
  names,
  with_geometries = FALSE,
  like = TRUE,
  fuzzy = FALSE
)
}
\arguments{
\item{names}{Terms to search in the Marine Regions Gazetteer given as a vector}

\item{with_geometries}{Logical. Add geometries to the result data frame? Default = FALSE}

\item{like}{Logical. Add a '%'-sign before and after the name? (SQL LIKE function). Default = TRUE}

\item{fuzzy}{Logical. Use Levenshtein query to find nearest matches? Default = FALSE}
}
\value{
Tibble with Gazetteer records.
}
\description{
Retrieve all Records of the [Marine Gazetteer](https://marineregions.org/gazetteer.php) that contain one or more parameters in `names` in their `preferredGazetteerName`.
}
\examples{
eez <- c("Belgian Exclusive Economic Zone", "Dutch Exclusive Economic Zone")
mr_gaz_records_by_names(eez, with_geometries = TRUE)
}
