% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_mr_gaz_sources.R
\name{mr_gaz_records_by_source}
\alias{mr_gaz_records_by_source}
\title{Retrieve Gazetteer Records by Source}
\usage{
mr_gaz_records_by_source(source, with_geometries = FALSE)
}
\arguments{
\item{source}{A source from mr_gaz_sources()}

\item{with_geometries}{Logical. Add geometries to the result data frame? Default = FALSE}
}
\value{
A tibble with all Gazetteer records of the specified source
}
\description{
Retrieve Gazetteer Records by Source
}
\examples{
ecoregions <- mr_gaz_records_by_source("ICES Ecoregions")
ecoregions$preferredGazetteerName[4]
}
