% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_gaz_records.R
\name{mr_gaz_records_by_type}
\alias{mr_gaz_records_by_type}
\title{Retrieve Gazetteer Records by Placetype}
\usage{
mr_gaz_records_by_type(type, offset = 0)
}
\arguments{
\item{type}{The placetype of which records are retrieved. Must be one of the types in `mr_gaz_info("placetypes")` and class `character`. Case insensitive.}

\item{offset}{Start record number, in order to page through next batch of results.}
}
\value{
A tibble with all Gazetteer records of the specified placetype.
}
\description{
Every record in the [Marine Gazetteer](https://marineregions.org/gazetteer.php) has a placetype, e.g. `Sandbank` or `Marine Protected Area`.
All placetypes currently available in the gazetteer can be retrieved via `mr_gaz_info("placetypes")`.
}
\examples{
fao_subdivs <- mr_gaz_records_by_type("FAO Subdivisions")
fao_subdivs$preferredGazetteerName[1]
# [1] "Northern Alboran Sea"
}
