% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mr_gaz_relations.R
\name{mr_gaz_relations_by_MRGID}
\alias{mr_gaz_relations_by_MRGID}
\title{Retrieve Gazetter Relations by MRGID}
\usage{
mr_gaz_relations_by_MRGID(
  mrgid,
  with_geometries = FALSE,
  direction = "both",
  type = "all"
)
}
\arguments{
\item{mrgid}{The [Marine Regions Geographic IDentifier](https://marineregions.org/mrgid.php).}

\item{with_geometries}{Logical. Add geometries to the result data frame? Default = FALSE}

\item{direction}{The hierarchical structure. Must be one of `c("upper", "lower", "both")`. `"upper"` lists all parents of the record. `"lower"` lists all childs of the record. `"both"` lists parents and childs of the record.}

\item{type}{must be one of `c("partof", "partlypartof", "adjacentto", "similarto", "administrativepartof", "influencedby", "all")`. Explanations of the `types` [here](https://marineregions.org/ontology/documentation.html) in chapter *Object Properties*.}
}
\value{
a `tibble` with all relations for the given `MRGID`.
}
\description{
Gazetteer Records that are related to a given input MRGID. Relationships can be parents (`upper`), children (`lower`) or both.
Geographic types can also be specified, for example `partof` and `adjacentto`.
}
\examples{
mariana_trench <- mr_gaz_records_by_names("Mariana Trench")
mariana_trench_mrgid <- mariana_trench$MRGID

mariana_trench_relations <- mr_gaz_relations_by_MRGID(mariana_trench_mrgid)
mariana_trench_relations$preferredGazetteerName
}
