% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregions2-package.R
\docType{package}
\name{mregions2-package}
\alias{mregions2}
\alias{mregions2-package}
\title{mregions2: Access Data from Marine Regions: Gazetteer & Data Products}
\description{
Explore and retrieve marine spatial data from the Marine Regions Gazetteer \url{https://marineregions.org/gazetteer.php?p=webservices} and the Marine Regions Data Products \url{https://marineregions.org/webservices.php}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/mregions2}
  \item \url{https://docs.ropensci.org/mregions2/}
  \item Report bugs at \url{https://github.com/ropensci/mregions2/issues}
}

}
\author{
\strong{Maintainer}: Salvador Jesús Fernández Bejarano \email{salvador.fernandez@vliz.be}

Authors:
\itemize{
  \item Lotte Pohl \email{lotte.pohl@vliz.be}
}

Other contributors:
\itemize{
  \item Julia Gustavsen [reviewer]
  \item Muralidhar M.A. [reviewer]
  \item Sheila M. Saia [reviewer]
}

}
\keyword{internal}
