% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregions2-package.R
\docType{package}
\name{mregions2-package}
\alias{mregions2}
\alias{mregions2-package}
\title{mregions2: Access Data from Marineregions.org: Gazetteer & Data Products}
\description{
Explore and retrieve marine geospatial data from the Marine Regions Gazetteer \url{https://marineregions.org/gazetteer.php?p=webservices} and the Marine Regions Data Products \url{https://marineregions.org/webservices.php}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/mregions2}
  \item \url{https://docs.ropensci.org/mregions2/}
  \item Report bugs at \url{https://github.com/ropensci/mregions2/issues}
}

}
\author{
\strong{Maintainer}: Salvador Jesús Fernández Bejarano \email{salvador.fernandez@vliz.be} (\href{https://orcid.org/0000-0003-0535-7677}{ORCID}) (salvafern)

Authors:
\itemize{
  \item Lotte Pohl \email{lotte.pohl@imbrsea.eu} (\href{https://orcid.org/0000-0002-7607-7018}{ORCID}) (lottepohl)
}

Other contributors:
\itemize{
  \item Julia Gustavsen [reviewer]
  \item Muralidhar M.A. [reviewer]
  \item Sheila M. Saia [reviewer]
  \item LifeWatch Belgium (lifewatch.be) [funder]
}

}
\keyword{internal}
