% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrp_list}
\alias{mrp_list}
\title{Available data products at Marine Regions}
\format{
\subsection{\code{mrp_list}}{

A data frame with 21 rows and 7 columns:
\describe{
\item{title}{Data product name}
\item{namespace}{Workspace in geoserver}
\item{layer}{Identifier of the data product. Use in \code{\link[=mrp_get]{mrp_get()}}}
\item{license}{terms of use of the data products}
\item{citation}{preferred citation of the data products}
\item{doi}{ISO 26324 \href{https://www.doi.org/}{Digital Object Identifier})}
\item{imis}{Url of the data products in the \href{https://vliz.be/en/imis}{Integrated Marine Information System (IMIS)}}
\item{abstract}{Description of the data product}
}
}
}
\source{
\url{https://marineregions.org/sources.php} \url{https://marineregions.org/eezmethodology.php}
}
\usage{
mrp_list
}
\description{
A data frame including the name, abstract and
some other relevant about each data product in Marine Regions.
}
\details{
Direct downloads are also available at: \url{https://marineregions.org/downloads.php}
}
\examples{
mrp_list
}
\keyword{datasets}
